backend/app/requirements.txt
django
djangorestframework
django-filter 
psycopg2>=2.8
django-cors-headers
androguard